+++ internal/venice-common/src/test/java/com/linkedin/venice/utils/TestStoreUtils.java
@@ -10,12 +10,12 @@ import org.testng.annotations.Test;
 public class TestStoreUtils {
   @Test
   public void testGetMaxRecordSizeBytes() {
-    long maxRecordSizeBytes = 1000L;
+    int maxRecordSizeBytes = 1000;
     Assert.assertEquals(
-        StoreUtils.getMaxRecordSizeBytes(true, -1L),
+        StoreUtils.getMaxRecordSizeBytes(true, -1),
         StoreUtils.DEFAULT_MAX_RECORD_SIZE_BYTES_WITH_CHUNKING);
     Assert.assertEquals(StoreUtils.getMaxRecordSizeBytes(true, maxRecordSizeBytes), maxRecordSizeBytes);
-    Assert.assertEquals(StoreUtils.getMaxRecordSizeBytes(false, -1L), -1L);
+    Assert.assertEquals(StoreUtils.getMaxRecordSizeBytes(false, -1), -1);
     Assert.assertEquals(StoreUtils.getMaxRecordSizeBytes(false, maxRecordSizeBytes), maxRecordSizeBytes);
   }
 }
